requirements.txt
aiohttp==3.9.5
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.9.1
asgiref==3.3.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
autobahn==20.12.3
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.11.1
bip32utils==0.3.post4
bitarray==2.9.2
blinker==1.8.2
boto3==1.34.117
botocore==1.34.117
bs4==0.0.1
cachetools==4.1.1
certifi==2024.6.2
cffi==1.14.4
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cloudinary==1.26.0
constantly==15.1.0
cryptography==3.3.1
cryptos==1.36
cytoolz==0.12.2
daphne==2.3.0
decorator==4.4.2
defusedxml==0.7.1
Django==2.2
django-allauth==0.51.0
django-amazon-ses==4.0.0
django-axes==5.9.0
django-betterforms==1.2
django-ckeditor==6.0.0
django-cors-headers==2.2.0
django-crispy-forms==1.13.0
django-crontab==0.7.1
django-filter==21.1
django-ipware==3.0.2
django-js-asset==1.2.2
django-simple-history==3.0.0
django-storages==1.12.3
django-tables2==2.4.1
django-user-agents==0.4.0
django-widget-tweaks==1.4.11
django-widgets-improved==1.5.0
djangoajax==2.3.5
djangorestframework==3.13.1
ecdsa==0.19.0
eth-abi==5.1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
exceptiongroup==1.2.1
flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==17.5.0
ipdb==0.13.4
ipfshttpclient==0.8.0a2
ipinfo==3.0.0
ipinfo-django==1.1.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.17.2
jinja2==3.1.4
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.3.0
MarkupSafe==2.1.5
mnemonic==0.21
msgpack==0.6.2
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.4
mysqlclient==2.2.4
netaddr==0.10.1
nltk==3.6.7
numpy==1.19.5
oauthlib==3.2.2
olefile==0.46
outcome==1.3.0.post0
package-name==0.1
pandas==1.1.5
parsimonious==0.10.0
parso==0.7.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.8
protobuf==5.29.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==1.4.0
pycointools==0.0.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.19.1
pydantic==2.10.5
pydantic-core==2.27.2
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==2.9.0
PyMySQL==1.1.0
pyOpenSSL==20.0.1
pyotp==2.4.1
pyrsistent==0.20.0
PySocks==1.7.1
python-bitcoinrpc==1.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2020.1
pyunormalize==16.0.0
referencing==0.32.0
regex==2022.6.2
requests==2.32.3
requests-oauthlib==1.3.1
rlp==1.2.0
rosetta==0.3
rpds-py==0.16.2
s3transfer==0.10.1
scipy==1.5.4
selenium==4.21.0
service-identity==18.1.0
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
toml==0.10.2
toolz==0.12.0
tqdm==4.64.0
traitlets==4.3.3
trio==0.25.1
trio-websocket==0.11.1
twilio==6.62.1
Twisted==20.3.0
txaio==20.12.1
types-requests==2.32.0.20241016
typing-extensions==4.12.1
tzlocal==5.2
ua-parser==0.10.0
urllib3==2.2.3
user-agents==2.2.0
varint==1.0.2
wcwidth==0.2.5
web3==6.0.0
websockets==13.1
werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.6.0
zope.interface==5.2.0
